[S07P22D102-29] fix : qr to Long
qr to Long

+++ AOS/InDive/app/src/main/java/com/ssafy/indive/view/home/HomeFragment.kt
@@ -71,14 +71,14 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 
     private fun initRecentMusic() {
 
-        homeViewModel.getMusics(null, null, null,"latest", null,null,null)
+        homeViewModel.getMusics(null, null, null, "latest", null, null, null)
         binding.rvRecentMusic.adapter = RecentMusicAdapter(playListener)
 
     }
 
     private fun initPopularMusic() {
 
-        homeViewModel.getMusics(null, null, null, "popular", null,0,4)
+        homeViewModel.getMusics(null, null, null, "popular", null, 0, 4)
 
         val moreListener: (MusicDetailResponse) -> (Unit) = {
             MoreDialogFragment(object : MoreDialogFragment.MoreDialogClickListener {
@@ -127,7 +127,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 
     fun scanQRSuccess() {
         val action =
-            HomeFragmentDirections.actionHomeFragmentToDonateFragment(MainActivity.successQRScanMsg)
+            HomeFragmentDirections.actionHomeFragmentToDonateFragment(MainActivity.successQRScanMsg.toLong())
         MainActivity.successQRScanMsg = ""
         findNavController().navigate(action)
     }
